+++ R/get_inboveg_layer_qualifier.R
@@ -1,11 +1,11 @@
-#' @title Query layer information of recordings (relevé) from INBOVEG
+#' @title Query layer qualifier information of recordings (relevé) from INBOVEG
 #'
 #' @description This function queries the INBOVEG database for layer
-#' information on recordings  for one or more surveys.
+#' qualifier information on recordings  for one or more surveys.
 #'
 #' @param survey_name A character string or a character vector, depending on
 #' multiple parameter, giving the name or names of the survey(s) for which you
-#' want to extract recordings information. If missing, all surveys are returned.
+#' want to extract information. If missing, all surveys are returned.
 #' @param connection dbconnection with the database 'Cydonia'
 #' on the inbo-sql07-prd server
 #' @param multiple If TRUE, survey_name can take a character vector with
@@ -14,8 +14,9 @@
 #' include wildcards to allow partial matches
 #'
 #' @return A dataframe with variables Name (of the survey), RecordingGivid
-#' (unique Id), UserReference, LayerCode, LayerDescription, CoverCode,
-#' Coverpercentage and Mean hight (cm)
+#' (unique Id), UserReference, LayerCode, LayerDescription, QualifierCode,
+#' Qualifier Description, Elucidation and 'NotSure' in case the qualifier is
+#' doubtful, CoverCode and Cover percentage
 #'
 #' @importFrom glue glue_sql
 #' @importFrom DBI dbGetQuery
@@ -28,19 +29,21 @@
 #' library(inbodb)
 #' con <- connect_inbo_dbase("D0010_00_Cydonia")
 #'
-#' # get the layer information from one survey
-#' layerinfo_heischraal2012 <- get_inboveg_layerinfo(con, survey_name =
-#' "MILKLIM_Heischraal2012")
+#' # get the layer qualifiers from one survey
+#' layerqualifiers_Gagealutea <-
+#'     get_inboveg_layer_qualifier(con, survey_name = "GageaLutea_1980")
 #'
 #' # get all layer qualifiers from MILKLIM surveys (partial matching)
-#' layerinfo_milkim <- get_inboveg_layerinfo(con, survey_name = "%MILKLIM%")
+#' layerqualifiers_milkim <-
+#'   get_inboveg_layer_qualifier(con, survey_name = "%MILKLIM%")
 #'
 #' # get layer qualifiers from several specific surveys
-#' layerinfo_severalsurveys <- get_inboveg_layerinfo(con, survey_name =
-#' c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE)
+#' layerqualifiers_severalsurveys <- get_inboveg_layer_qualifier(con,
+#'   survey_name = c("MILKLIM_Overstroming", "NICHE Vlaanderen"),
+#'   multiple = TRUE)
 #'
 #' # get all layer qualifiers of all surveys
-#' all_layerinfo <- get_inboveg_layerinfo(con)
+#' alllayerqualifiers <- get_inboveg_layer_qualifier(con)
 #'
 #' # Close the connection when done
 #' dbDisconnect(con)
@@ -49,9 +52,9 @@
 #'
 
 
-get_inboveg_layerinfo <- function(connection,
-                                       survey_name,
-                                       multiple = FALSE) {
+get_inboveg_layer_qualifier <- function(connection,
+                                   survey_name,
+                                   multiple = FALSE) {
 
   assert_that(inherits(connection, what = "Microsoft SQL Server"),
               msg = "Not a connection object to database.")
@@ -79,27 +82,22 @@ common_part <-
     , ivRecording.UserReference
     , ivRLLayer.LayerCode
     , ftAGV.Description as LayerDescription
-    , ivRLLayer.CoverCode
-    , ftAGV_01.Description as Percentage
+    , ivRLQualifier.QualifierCode
+    , ivRLQualifier.Elucidation
+    , ivRLQualifier.NotSure
   FROM ivRecording
     INNER JOIN ivSurvey ivS on ivS.Id = ivRecording.SurveyId
-    INNER JOIN  ivRLLayer on ivRLLayer.RecordingID = ivRecording.Id
-      INNER JOIN ivRLResources
-          on ivRLResources.ResourceGIVID = ivRLLayer.LayerResource
-        LEFT JOIN [syno].[Futon_dbo_ftActionGroupValues] ftAGV
+    LEFT JOIN  ivRLLayer on ivRLLayer.RecordingID = ivRecording.Id
+    LEFT JOIN ivRLQualifier ON ivRLLayer.ID = ivRLQualifier.LayerID
+    LEFT JOIN ivRLResources
+        on ivRLResources.ResourceGIVID = ivRLLayer.LayerResource
+    LEFT JOIN [syno].[Futon_dbo_ftActionGroupValues] ftAGV
         ON ivRLResources.ListName = ftAGV.ListName COLLATE Latin1_General_CI_AI
         AND ivRLResources.ActionGroup = ftAGV.ActionGroup
             COLLATE Latin1_General_CI_AI
-        AND ivRLLayer.LayerCode = ftAGV.code COLLATE Latin1_General_CI_AI
-      INNER JOIN ivRLResources ivRLR_01
-        on ivRLR_01.ResourceGIVID = ivRLLayer.CoverResource
-          LEFT JOIN [syno].[Futon_dbo_ftActionGroupValues] ftAGV_01
-            ON ivRLR_01.ListName = ftAGV_01.ListName
-                COLLATE Latin1_General_CI_AI
-          AND ivRLR_01.ActionGroup = ftAGV_01.ActionGroup
-              COLLATE Latin1_General_CI_AI
-          AND ivRLLayer.CoverCode = ftAGV_01.code COLLATE Latin1_General_CI_AI
+        AND ivRLLayer.LayerCode = ftAGV.Code COLLATE Latin1_General_CI_AI
   WHERE 1 = 1"
+
 
   if (!multiple) {
     sql_statement <- glue_sql(common_part,
@@ -116,7 +114,7 @@ common_part <-
 
   sql_statement <- glue_sql(
     sql_statement,
-    "ORDER BY ivRecording.RecordingGivid, ivRLLayer.LayerCode",
+   "ORDER BY ivRecording.RecordingGivid, ivRLLayer.LayerCode",
     .con = connection)
 
   query_result <- dbGetQuery(connection, sql_statement)