+++ R/get_inboveg_recording.R
@@ -6,15 +6,16 @@
 #' or in combination with the unique ID (recordingGIVID) or user reference
 #'
 #' @param user_reference A character string or a character vector giving the
-#' name of a recording for which you want to extract releve information. If
-#' missing, all user-references are returned.
+#' name of a recording for which you want to extract releve information.
+#' As default (user_reference = "%") all user-references are returned.
 #' @param recording_givid A character string or a character vector giving
 #' the unique id of a recording for which you want to extract releve
 #' information.
-#' If missing, all recording_givids are returned.
+#' As default (recording_givids = "%") all recording_givids are returned.
 #' @param survey_name A character string or a character vector, depending on
 #' multiple parameter, giving the name or names of the survey(s) for which you
-#' want to extract releve information. If missing, all surveys are returned.
+#' want to extract releve information. As default (surey_name = "%") all surveys
+#' are returned.
 #' @param connection dbconnection with the database 'Cydonia'
 #' on the inbo-sql07-prd server
 #' @param collect If FALSE (the default), a remote tbl object is returned.
@@ -22,9 +23,10 @@
 #' the query is not brought into memory. If TRUE the full result of the query is
 #' collected (fetched) from the database and brought into memory of the working
 #' environment.
-#' @param multiple If TRUE, survey_name, user_reference or recording_givid can
-#' take a character vector with multiple survey names that must match exactly.
-#' If FALSE (the default), survey_name , user_reference or recording_givid must
+#' @param multiple If TRUE survey_name, user_reference or recording_givid can
+#' take a character vector with multiple survey names, user references or
+#' recording_givids that must match exactly.
+#' If FALSE (the default) survey_name , user_reference or recording_givid must
 #' be a single character string (one survey name, or one user_reference or one
 #' recording_givid). Only survey_name can include wildcarts to allow partial
 #' matches.
@@ -48,74 +50,44 @@
 #' con <- connect_inbo_dbase("D0010_00_Cydonia")
 #'
 #' # get the recordings from one survey and collect the data
-#' recording_heischraal2012 <- get_inboveg_recordings(con, survey_name =
+#' recording_heischraal2012 <- get_inboveg_recording(con, survey_name =
 #' "MILKLIM_Heischraal2012", collect = TRUE)
 #'
 #' # get all recordings from MILKLIM surveys (partial matching), don't collect
-#' recording_milkim <- get_inboveg_recordings(con, survey_name = "%MILKLIM%",
+#' recording_milkim <- get_inboveg_recording(con, survey_name = "%MILKLIM%",
 #' collect = FALSE)
 #'
 #' # get recordings from several specific surveys
-#' recording_severalsurveys <- get_inboveg_recordings(con, survey_name =
+#' recording_severalsurveys <- get_inboveg_recording(con, survey_name =
 #' c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE,
 #' collect = TRUE)
 #'
 #' # get recordings from several specific recordinggivid
-#' recording_severalgivids <- get_inboveg_recordings(con,
+#' recording_severalgivids <- get_inboveg_recording(con,
 #' recording_givid = c("IV2012081609450300","IV2012081610204607"),
 #' multiple = TRUE, collect = TRUE)
 #'
 #' # get all recordings of all surveys,  don't collect the data
-#' allrecordings <- get_inboveg_recordings(con)
+#' allrecordings <- get_inboveg_recording(con)
 #'
 #' # Close the connection when done
 #' dbDisconnect(con)
 #' rm(con)
 #' }
 
-get_inboveg_recordings <- function(
+get_inboveg_recording <- function(
   connection,
-  survey_name,
-  user_reference,
-  recording_givid,
+  survey_name = "%",
+  user_reference = "%",
+  recording_givid = "%",
   collect = FALSE,
   multiple = FALSE) {
 
   assert_that(inherits(connection, what = "Microsoft SQL Server"),
               msg = "Not a connection object to database.")
-
-  if (!multiple) {
-    if (missing(survey_name)) {
-      survey_name <- "%"
-    }
-    assert_that(is.character(survey_name))
-    if (missing(user_reference)) {
-      user_reference <- "%"
-    }
-    assert_that(is.character(user_reference))
-    if (missing(recording_givid)) {
-      recording_givid <- "%"
-    }
-    assert_that(is.character(recording_givid))
-  } else {
-    if (missing(survey_name) & missing(recording_givid) &
-        missing(user_reference)) {
-      stop("Please provide one or more values to either survey_name or to user_reference or recording_givid when multiple = TRUE") #nolint
-    }
-    if (missing(survey_name) & !missing(user_reference)) {
-      warning("The same user_reference might have been used in different surveys. Consider also providing one or more values to survey_name when multiple = TRUE") #nolint
-    }
-    if (!missing(survey_name)) {
-      assert_that(is.vector(survey_name, mode = "character"))
-    }
-    if (!missing(user_reference)) {
-      assert_that(is.vector(user_reference, mode = "character"))
-    }
-    if (!missing(recording_givid)) {
-      assert_that(is.vector(recording_givid, mode = "character"))
-    }
-  }
-
+  assert_that(is.character(survey_name))
+  assert_that(is.character(user_reference))
+  assert_that(is.character(recording_givid))
 
   common_part <- "SELECT ivS.Name
   , ivR.[RecordingGivid]
@@ -174,6 +146,14 @@ get_inboveg_recordings <- function(
   AND ivRL_Iden.Preferred = 1"
 
   if (!multiple) {
+
+    assert_that(length(survey_name) == 1,
+                msg = "If you want to give multiple survey names, please change multiple  into TRUE") #nolint
+    assert_that(length(recording_givid) == 1,
+                msg = "If you want to give multiple recording givid, please change multiple  into TRUE") #nolint
+    assert_that(length(user_reference) == 1,
+                msg = "If you want to give multiple user reference, please change multiple into TRUE") #nolint
+
     sql_statement <- glue_sql(common_part,
                               "AND ivS.Name LIKE {survey_name}
                               AND ivR.[RecordingGivid] LIKE {recording_givid}
@@ -184,8 +164,9 @@ get_inboveg_recordings <- function(
                               .con = connection)
 
   } else {
-    if (!missing(survey_name) & !missing(user_reference) &
-        !missing(recording_givid)) {
+    if (!(length(survey_name) == 1 & survey_name[1] == "%") &
+        !(length(user_reference) == 1 & user_reference[1] == "%") &
+        !(length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivS.Name IN ({survey_name*})
                                 AND (ivR.[RecordingGivid] IN
@@ -196,65 +177,65 @@ get_inboveg_recordings <- function(
                                 recording_givid = recording_givid,
                                 .con = connection)
     }
-    if (missing(survey_name) & !missing(user_reference) &
-        !missing(recording_givid)) {
+    if ((length(survey_name) == 1 & survey_name[1] == "%") &
+        !(length(user_reference) == 1 & user_reference[1] == "%") &
+        !(length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND (ivR.[RecordingGivid] IN
                                     ({recording_givid*})
                                 OR ivR.UserReference IN ({user_reference*}))",
-                                survey_name = survey_name,
                                 user_reference = user_reference,
                                 recording_givid = recording_givid,
                                 .con = connection)
     }
-    if (missing(survey_name) & missing(user_reference) &
-        !missing(recording_givid)) {
+    if ((length(survey_name) == 1 & survey_name[1] == "%") &
+        (length(user_reference) == 1 & user_reference[1] == "%") &
+        !(length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivR.[RecordingGivid] IN
                                    ({recording_givid*})",
-                                survey_name = survey_name,
-                                user_reference = user_reference,
                                 recording_givid = recording_givid,
                                 .con = connection)
     }
-    if (missing(survey_name) & !missing(user_reference) &
-        missing(recording_givid)) {
+    if ((length(survey_name) == 1 & survey_name[1] == "%") &
+        !(length(user_reference) == 1 & user_reference[1] == "%") &
+        (length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivR.UserReference IN ({user_reference*})",
-                                survey_name = survey_name,
                                 user_reference = user_reference,
-                                recording_givid = recording_givid,
                                 .con = connection)
     }
-    if (!missing(survey_name) & missing(user_reference) &
-        missing(recording_givid)) {
+    if (!(length(survey_name) == 1 & survey_name[1] == "%") &
+        (length(user_reference) == 1 & user_reference[1] == "%") &
+        (length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivS.Name IN ({survey_name*})",
                                 survey_name = survey_name,
-                                user_reference = user_reference,
-                                recording_givid = recording_givid,
-                                .con = connection)
+                              .con = connection)
     }
-    if (!missing(survey_name) & missing(user_reference) &
-        !missing(recording_givid)) {
+    if (!(length(survey_name) == 1 & survey_name[1] == "%") &
+        (length(user_reference) == 1 & user_reference[1] == "%") &
+        !(length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivS.Name IN ({survey_name*})
                                 AND ivR.[RecordingGivid] IN
                                     ({recording_givid*})",
                                 survey_name = survey_name,
-                                user_reference = user_reference,
                                 recording_givid = recording_givid,
                                 .con = connection)
     }
-    if (!missing(survey_name) & !missing(user_reference) &
-        missing(recording_givid)) {
+    if (!(length(survey_name) == 1 & survey_name[1] == "%") &
+        !(length(user_reference) == 1 & user_reference[1] == "%") &
+        (length(recording_givid) == 1 & recording_givid[1] == "%")) {
       sql_statement <- glue_sql(common_part,
                                 "AND ivS.Name IN ({survey_name*})
                                 AND ivR.UserReference IN ({user_reference*})",
                                 survey_name = survey_name,
                                 user_reference = user_reference,
-                                recording_givid = recording_givid,
                                 .con = connection)
+    }
+    if (!exists("sql_statement")) {
+      stop("When choosing multiple = TRUE, please add a (vector of) value(s) for at least one of the following variables: survey_name, user_reference or recording_givid.") #nolint
     }
 
   }
@@ -265,4 +246,21 @@ get_inboveg_recordings <- function(
     query_result <- collect(query_result)
   }
   return(query_result)
+}
+
+get_inboveg_recordings <- function(
+  connection,
+  survey_name = "%",
+  user_reference = "%",
+  recording_givid = "%",
+  collect = FALSE,
+  multiple = FALSE) {
+
+  .Deprecated("get_inboveg_recording")
+  get_inboveg_recording(connection = connection,
+                        survey_name = survey_name,
+                        user_reference = user_reference,
+                        recording_givid = recording_givid,
+                        collect = collect,
+                        multiple = multiple)
 }