+++ NEWS.md
@@ -1,3 +1,7 @@
 # inbodb 0.0.2
 
+* New functions `get_inboveg_layer_cover()` and `get_inboveg_layer_qualifier()`.
+* Renamed functions `get_inboveg_qualifiers()` as `get_inboveg_qualifier()`,
+  `get_inboveg_recordings()` as `get_inboveg_recording()` and
+  `get_inboveg_relation()` as `get_inboveg_relation_recording()`.
 * Added a `NEWS.md` file to track changes to the package.